frontend: one link vocabulary, hash routing in its own module, footer held equal to the book's
Self-review of the branch, read as one diff.

The link set was split: `ProjectLink` in controls.rs, `Chapter` in
guide.rs, and two sum types over them (`GuideLink`, `FooterLink`) with
the same render shape. `links.rs` now owns all of it: `ProjectLink`,
`Chapter`, and one `SiteLink` (page, project resource, chapter) that
every surface renders in its own class. controls.rs keeps the shell's
own vocabulary and shrinks back under its former size.

app.rs had crossed 1000 lines. The hash router (parse, print, history
writes, the extension override) is `app/routes.rs`, with its round-trip
test; `shell_route_fragment` prints `ShellPage::slug`, which is also the
document key, so a page's name has one source and the test holds parser
and printer inverse.

`text_input` and `text_input_with_suggestions` repeated the field; both
are `text_field` with an optional datalist. The public seed endpoint is
a literal macro spliced into the guide's listings with `concat!`, so the
address is written once.

The book's footer is a second copy of the frontend's, in Handlebars. A
test in footer.rs includes `docs/theme/index.hbs` and asserts every
statement and destination appears there, so the copies cannot drift
unnoticed. The book's Docs link is absolute for the same reason. The
theme menu offers the two themes rings.css restyles; a theme stored by
an earlier build falls back to the default, and the popup anchors to the
controls it belongs to.
