+++ docs/theme/index.hbs
@@ -104,7 +104,13 @@
             const default_theme = window.matchMedia("(prefers-color-scheme: dark)").matches ? default_dark_theme : default_light_theme;
             let theme;
             try { theme = localStorage.getItem('mdbook-theme'); } catch(e) { }
-            if (theme === null || theme === undefined) { theme = default_theme; }
+            // Only the two themes rings.css restyles exist here; a theme stored by an earlier
+            // build of the book falls back to the default, and its record is dropped so the
+            // theme menu never looks up a button that is not there.
+            if (theme !== 'light' && theme !== 'navy') {
+                theme = default_theme;
+                try { localStorage.removeItem('mdbook-theme'); } catch(e) { }
+            }
             const html = document.documentElement;
             html.classList.remove('{{ default_theme }}')
             html.classList.add(theme);
@@ -169,13 +175,12 @@
                             <button id="mdbook-theme-toggle" class="icon-button" type="button" title="Change theme" aria-label="Change theme" aria-haspopup="true" aria-expanded="false" aria-controls="mdbook-theme-list">
                                 {{fa "solid" "paintbrush"}}
                             </button>
+                            <!-- Only the two themes rings.css restyles are offered; `navy` is the
+                                 book's name for the dark one. -->
                             <ul id="mdbook-theme-list" class="theme-popup" aria-label="Themes" role="menu">
                                 <li role="none"><button role="menuitem" class="theme" id="mdbook-theme-default_theme">Auto</button></li>
                                 <li role="none"><button role="menuitem" class="theme" id="mdbook-theme-light">Light</button></li>
-                                <li role="none"><button role="menuitem" class="theme" id="mdbook-theme-rust">Rust</button></li>
-                                <li role="none"><button role="menuitem" class="theme" id="mdbook-theme-coal">Coal</button></li>
-                                <li role="none"><button role="menuitem" class="theme" id="mdbook-theme-navy">Navy</button></li>
-                                <li role="none"><button role="menuitem" class="theme" id="mdbook-theme-ayu">Ayu</button></li>
+                                <li role="none"><button role="menuitem" class="theme" id="mdbook-theme-navy">Dark</button></li>
                             </ul>
                             {{#if search_enabled}}
                             <button id="mdbook-search-toggle" class="icon-button" type="button" title="Search (`/`)" aria-label="Toggle Searchbar" aria-expanded="false" aria-keyshortcuts="/ s" aria-controls="mdbook-searchbar">
@@ -279,7 +284,7 @@
                         <nav class="rings-footer-column" aria-label="Developers">
                             <h3>Developers</h3>
                             <ul>
-                                <li><a href="{{ path_to_root }}">Docs</a></li>
+                                <li><a href="https://rings.rs/docs/">Docs</a></li>
                                 <li><a href="https://github.com/RingsNetwork/rings">GitHub</a></li>
                                 <li><a href="https://github.com/RingsNetwork/rings/releases">Releases</a></li>
                                 <li><a href="https://crates.io/crates/rings-node">crates.io</a></li>
