+++ frontend/src/footer.rs
@@ -4,11 +4,16 @@
 //! console is an application screen rather than a document and does not carry it, and below
 //! the mobile breakpoint the fixed bottom navigation bar owns the same space, so the stylesheet
 //! hides it there (`styles::theme::responsive`).
+//!
+//! The documentation book renders the same footer from `docs/theme/index.hbs`. The two are
+//! written in different languages, so they are two copies; the test at the bottom of this
+//! module holds them equal.
 
 use yew::prelude::*;
 
-use crate::controls::ProjectLink;
 use crate::controls::ShellPage;
+use crate::links::ProjectLink;
+use crate::links::SiteLink;
 
 pub(crate) fn site_footer(navigate_page: Callback<ShellPage>) -> Html {
     html! {
@@ -22,9 +27,7 @@ pub(crate) fn site_footer(navigate_page: Callback<ShellPage>) -> Html {
                         decoding="async"
                     />
                     <strong>{ "Rings Network" }</strong>
-                    <p>
-                        { "A peer-to-peer network for the sovereign age. Browser tabs and native daemons join one overlay, find each other by DID, and talk over direct WebRTC datachannels routed by a Chord DHT." }
-                    </p>
+                    <p>{ BRAND_STATEMENT }</p>
                 </div>
                 { for COLUMNS.iter().map(|column| column.render(&navigate_page)) }
             </div>
@@ -44,26 +47,12 @@ fn current_year() -> u32 {
     js_sys::Date::new_0().get_full_year()
 }
 
-/// A footer destination: a page of this shell or a project resource.
-#[derive(Clone, Copy)]
-enum FooterLink {
-    Page(ShellPage),
-    Project(ProjectLink),
-}
-
-impl FooterLink {
-    fn render(self, navigate_page: &Callback<ShellPage>) -> Html {
-        match self {
-            Self::Page(page) => page.button("site-footer-link", false, navigate_page),
-            Self::Project(link) => link.anchor("site-footer-link"),
-        }
-    }
-}
+const BRAND_STATEMENT: &str = "A peer-to-peer network for the sovereign age. Browser tabs and native daemons join one overlay, find each other by DID, and talk over direct WebRTC datachannels routed by a Chord DHT.";
 
 /// A footer column: a heading over its destinations.
 struct Column {
     title: &'static str,
-    links: &'static [FooterLink],
+    links: &'static [SiteLink],
 }
 
 impl Column {
@@ -72,7 +61,9 @@ impl Column {
             <nav class="site-footer-column" aria-label={self.title}>
                 <h3>{ self.title }</h3>
                 <ul>
-                    { for self.links.iter().map(|link| html! { <li>{ link.render(navigate_page) }</li> }) }
+                    { for self.links.iter().map(|link| html! {
+                        <li>{ link.render("site-footer-link", navigate_page) }</li>
+                    }) }
                 </ul>
             </nav>
         }
@@ -83,30 +74,30 @@ const COLUMNS: [Column; 3] = [
     Column {
         title: "Product",
         links: &[
-            FooterLink::Page(ShellPage::Home),
-            FooterLink::Page(ShellPage::Guide),
-            FooterLink::Page(ShellPage::Console),
+            SiteLink::Page(ShellPage::Home),
+            SiteLink::Page(ShellPage::Guide),
+            SiteLink::Page(ShellPage::Console),
         ],
     },
     Column {
         title: "Developers",
         links: &[
-            FooterLink::Project(ProjectLink::Docs),
-            FooterLink::Project(ProjectLink::Repository),
-            FooterLink::Project(ProjectLink::Releases),
-            FooterLink::Project(ProjectLink::CratesIo),
-            FooterLink::Project(ProjectLink::Npm),
-            FooterLink::Project(ProjectLink::LlmsTxt),
+            SiteLink::Project(ProjectLink::Docs),
+            SiteLink::Project(ProjectLink::Repository),
+            SiteLink::Project(ProjectLink::Releases),
+            SiteLink::Project(ProjectLink::CratesIo),
+            SiteLink::Project(ProjectLink::Npm),
+            SiteLink::Project(ProjectLink::LlmsTxt),
         ],
     },
     Column {
         title: "Project",
         links: &[
-            FooterLink::Project(ProjectLink::Whitepaper),
-            FooterLink::Project(ProjectLink::Security),
-            FooterLink::Project(ProjectLink::Roadmap),
-            FooterLink::Project(ProjectLink::Sponsor),
-            FooterLink::Project(ProjectLink::License),
+            SiteLink::Project(ProjectLink::Whitepaper),
+            SiteLink::Project(ProjectLink::Security),
+            SiteLink::Project(ProjectLink::Roadmap),
+            SiteLink::Project(ProjectLink::Sponsor),
+            SiteLink::Project(ProjectLink::License),
         ],
     },
 ];
@@ -150,3 +141,43 @@ const NOTICES: [Notice; 5] = [
         body: "rings.rs is a static site: no accounts, no analytics, no tracking cookies. Browser-node settings are kept in your browser's local storage.",
     },
 ];
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    /// The book's page template, which carries this footer in Handlebars.
+    const BOOK_TEMPLATE: &str = include_str!("../../docs/theme/index.hbs");
+
+    /// Law: every statement and every destination of this footer appears verbatim in the
+    /// book's footer, so the two copies cannot drift apart unnoticed.
+    #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test::wasm_bindgen_test)]
+    #[cfg_attr(not(target_arch = "wasm32"), test)]
+    fn test_book_footer_carries_the_same_statements_and_destinations() {
+        assert!(BOOK_TEMPLATE.contains(BRAND_STATEMENT));
+        for notice in NOTICES.iter() {
+            assert!(
+                BOOK_TEMPLATE.contains(notice.title) && BOOK_TEMPLATE.contains(notice.body),
+                "the book footer lacks the {} notice",
+                notice.title
+            );
+        }
+        for column in COLUMNS.iter() {
+            assert!(BOOK_TEMPLATE.contains(column.title));
+            for link in column.links.iter() {
+                let href = match link {
+                    SiteLink::Project(project) => project.href().to_string(),
+                    SiteLink::Chapter(chapter) => chapter.href().to_string(),
+                    // A page is reached from the book through the site's hash route, the
+                    // landing page through the site root.
+                    SiteLink::Page(ShellPage::Home) => "https://rings.rs/\"".to_string(),
+                    SiteLink::Page(page) => format!("https://rings.rs/#{}", page.slug()),
+                };
+                assert!(
+                    BOOK_TEMPLATE.contains(&href),
+                    "the book footer lacks {href}"
+                );
+            }
+        }
+    }
+}
